rename blu_common_ui.ts
