+++ packages/frontend/src/scripts/sims/blu/blu_common_ui.ts
